requirements.txt
# Flow Matching RD图 Sim2Real 项目依赖

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# 数据处理
numpy>=1.24.0
pillow>=9.0.0

# 配置文件
pyyaml>=6.0

# 可视化
matplotlib>=3.7.0
tensorboard>=2.13.0

# 进度条
tqdm>=4.65.0

# 图像处理（可选，用于感知损失）
# torchvision已包含，无需额外安装

# 工具
pathlib2>=2.3.7; python_version < '3.4'

